% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdisfdata.R
\docType{package}
\name{fdisfdata}
\alias{fdisfdata-package}
\alias{fdisfdata}
\title{Spatial datasets for fisheries data interoperability}
\description{
fdisfdata provides a set of spatial datasets for fisheries data interoperability
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fdiwg/fdisfdata}
  \item Report bugs at \url{https://github.com/fdiwg/fdisfdata/issues}
}

}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
